fixing some stuff, switching to baos bais

+++ iguana.corecontroller/src/main/java/org/aksw/iguana/cc/lang/impl/ThrowawayLanguageProcessor.java
@@ -1,6 +1,7 @@
 package org.aksw.iguana.cc.lang.impl;
 
 import org.aksw.iguana.cc.lang.AbstractLanguageProcessor;
+import org.aksw.iguana.commons.io.BigByteArrayOutputStream;
 import org.aksw.iguana.commons.streams.Streams;
 
 import java.io.ByteArrayOutputStream;
@@ -12,7 +13,7 @@ import java.util.concurrent.TimeoutException;
 public class ThrowawayLanguageProcessor extends AbstractLanguageProcessor {
 
     @Override
-    public int readResponse(InputStream inputStream, Instant startTime, Double timeOut, ByteArrayOutputStream responseBody) throws IOException, TimeoutException {
+    public long readResponse(InputStream inputStream, Instant startTime, Double timeOut, ByteArrayOutputStream responseBody) throws IOException, TimeoutException {
         return Streams.inputStream2Length(inputStream, startTime, timeOut);
     }
